+++ pkg/server/types.go
@@ -16,7 +16,6 @@ package server
 
 import (
 	"github.com/gardener/etcd-backup-restore/pkg/compressor"
-	"github.com/gardener/etcd-backup-restore/pkg/leaderelection"
 	brtypes "github.com/gardener/etcd-backup-restore/pkg/types"
 )
 
@@ -38,7 +37,7 @@ type BackupRestoreComponentConfig struct {
 	DefragmentationSchedule string                        `json:"defragmentationSchedule"`
 	EtcdProcessName         string                        `json:"etcdProcessName"`
 	HealthConfig            *brtypes.HealthConfig         `json:"healthConfig,omitempty"`
-	LeaderElectionConfig    *leaderelection.Config        `json:"leaderElectionConfig,omitempty"`
+	LeaderElectionConfig    *brtypes.Config               `json:"leaderElectionConfig,omitempty"`
 }
 
 // latestSnapshotMetadata holds snapshot details of latest full and delta snapshots
